correction boucle infinie quand GPS branché pendant une fonction bloquante (#8578)
git-svn-id: http://einstein/svn/firmware/FIR-00251-Common/trunk@19728 3e98269d-ff51-e449-abca-491a6567d830

+++ Software/CircularByteBuffer.c
@@ -420,8 +420,8 @@ IRC_Status_t CBB_Dump(circByteBuffer_t *circByteBuffer, uint32_t lineSize)
    }
 
 #ifdef ENABLE_PRINTF
-   uint32_t i;
-   uint32_t length = CBB_Length(circByteBuffer);
+   uint16_t i;
+   uint16_t length = CBB_Length(circByteBuffer);
    uint32_t baseAddr = 0;
 
    for (i = 0; i < length; i++)
